update create_prefix_df
dealing with NA's. Leave out the last activity in a trace, for which no next_time exists (because it is not followed by any other activity).
Assumption: the input trace/activity to predict on should not be considered complete. Providing an input means that you expect the input to be followed by another activity which is yet to be executed.